Add Images to Projects Give the a face, something to present them, something to judge them.
